Update service worker build version and refactor Summary component import
- Updated BUILD_VERSION in service-worker.js to ensure consistent cache versioning.
- Refactored import path for Summary component in page.tsx to use the canonical path.
- Adjusted layout in page.tsx for better structure and removed unnecessary nesting in the Summary card.
- Added "use client" directive at the top of Summary.tsx for client-side rendering.
